requirements.txt
playwright==1.*
beautifulsoup4==4.*
requests==2.*
python-dotenv==1.*
